FEAT: Implement dynamic form validation and organization in AddPetScreen
This commit introduces a new FormStateManager to handle dynamic form validation and state management within the AddPetScreen. It also refactors the component structure for better organization.

- Implemented a FormStateManager to manage form validation and state updates.
- Introduced a FieldStatesCollector utility for collecting and managing the states of form fields.
- Added an AddPetContent composable to hold the progress indicator and the form fields.
- Refactored form field components into a more organized structure, moving them into dedicated packages.
- Improved the organization of the codebase by refactoring the component structure and introducing new utils.
- Removed unnecessary recomposition of form fields.
- Added logic to conditionally enable the add button if all the fields are valid.
- Modified AddPetViewmodel to support the new validation logic.

- Created the files:
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/AddPetContent.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/utils/FormStateManager.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/utils/FieldStatesCollecter.kt`
- Created the folder:
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields`
- Modified the files:
`app/src/main/java/com/example/feedm/petsFeature/ui/viewmodel/AddPetViewmodel.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/AddPetScreen.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/DashBoardScreen.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/utils/FormItemsInteractionsHandler.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/Form.kt`
- Moved the files to new folders:
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/WeightField.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/SterilizationField.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/ProgressIndicator.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/AgeField.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/ActivityField.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/GoalField.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/FormField.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/SexField.kt`
`app/src/main/java/com/example/feedm/petsFeature/ui/view/screens/addPetScreen/components/addPetContentComponents/formFields/PetNameAndAnimalField.kt`
`app/src/main/java/com/example/feedm/core/navigation/Screens.kt`
