+++ app/src/main/java/com/example/feedm/core/navigation/Screens.kt
@@ -11,7 +11,6 @@ object AddPetScreen
 @Serializable
 data class AddMealScreen(val petId: Int? = null, val foodId: Int? = null)
 
-
 @Serializable
 data class AddFoodScreen(val origin: String? = null)
